+++ ui/address/contract/methods/form/__screenshots__/ContractMethodResultApi.pw.tsx_default_complex-success-1.png
